back-end/services/audio-watermarking-api/requirements.txt
fastapi
python-dotenv
uvicorn
audioseal
python-multipart
torch
torchaudio
numpy
soundfile

